requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
bitarray==3.1.1
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.1.0
click==8.1.7
cloup==3.0.5
colorama==0.4.6
cytoolz==1.0.1
decorator==5.1.1
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
frozenlist==1.5.0
glcontext==2.5.0
hexbytes==1.3.0
idna==3.10
isosurfaces==0.1.2
joblib==1.4.2
manim==0.18.1
ManimPango==0.5.0
mapbox-earcut==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
moderngl==5.10.0
moderngl-window==2.4.6
multidict==6.1.0
multipledispatch==1.0.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.82
pandas==2.2.3
parsimonious==0.10.0
pillow==10.3.0
propcache==0.3.0
pycairo==1.26.0
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyglet==2.0.15
Pygments==2.18.0
pyrr==0.10.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
pyunormalize==16.0.0
pywin32==309
regex==2024.11.6
requests==2.32.3
rich==13.7.1
rlp==4.1.0
scikit-learn==1.6.1
scipy==1.13.1
screeninfo==0.8.1
six==1.17.0
skia-pathops==0.8.0.post1
srt==3.5.3
svgelements==1.9.6
threadpoolctl==3.6.0
toolz==1.0.0
tqdm==4.66.4
types-requests==2.32.0.20250306
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
watchdog==4.0.1
web3==7.9.0
websockets==13.1
yarl==1.18.3
